chore: creating utility functions. chore: creating selectTodos selector
