+++ src/Components/ListTodos/todoList.css
@@ -3,14 +3,6 @@
   flex-wrap: wrap;
   justify-content: space-between;
   gap: 10px;
-
-  box-shadow: 7px 7px 15px rgba(55, 84, 170, 0.15),
-    -7px -7px 20px rgba(255, 255, 255, 1),
-    inset 0px 0px 4px rgba(255, 255, 255, 0.2),
-    inset 7px 7px 15px rgba(55, 84, 170, 0),
-    inset -7px -7px 20px rgba(255, 255, 255, 0),
-    0px 0px 4px rgba(255, 255, 255, 0);
-  transition: all 0.5s;
 }
 
 .todo {
